Factor const_state.rs, json_state.rs, root_state.rs out of shared.rs.
