+++ nanvm-lib/src/parser/mod.rs
@@ -3,6 +3,9 @@
 #![allow(clippy::module_inception)]
 pub mod analyzer;
 pub mod any_state;
+pub mod const_state;
+pub mod json_state;
 pub mod parser;
 pub mod path;
+pub mod root_state;
 pub mod shared;
