+++ nanvm-lib/src/parser/parser.rs
@@ -1,10 +1,10 @@
 use super::{
     any_state::{AnyResult, AnyState},
+    const_state::ConstState,
+    json_state::JsonState,
     path::{concat, split},
-    shared::{
-        ConstState, JsonElement, JsonState, ModuleCache, ParseError, ParseResult, ParsingStatus,
-        RootState, RootStatus,
-    },
+    root_state::{RootState, RootStatus},
+    shared::{JsonElement, ModuleCache, ParseError, ParseResult, ParsingStatus},
 };
 use crate::{
     common::default::default,
